refactor officer and incident pages
