--- frontend/compositions/officer-view/index.ts
@@ -1,6 +1,0 @@
-import OfficerHeader from "./officer-view-header"
-import OptionalOfficerInfo from "./optional-officer-info"
-import OfficerWorkHistory from "./officer-work-history"
-import OfficerAffiliations from "./officer-affiliations"
-
-export { OfficerHeader, OptionalOfficerInfo, OfficerWorkHistory, OfficerAffiliations }
