+++ frontend/pageStories/officer-view.stories.tsx
@@ -1,10 +1,10 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react"
-import OfficerView from "../pages/officer-view"
+import OfficerView from "../pages/officer/[id]"
 
 export default {
   title: "Pages/OfficerView"
 } as ComponentMeta<typeof OfficerView>
 
-const Template: ComponentStory<typeof OfficerView> = (args) => <OfficerView {...args} />
+const Template: ComponentStory<typeof OfficerView> = () => <OfficerView />
 
 export const Timothy = Template.bind({})
